Refactor Frontend Components
Moved around the frontend components of the project into their own subfolders to make navigating the project easier.

There are 4 new folders:

- `frontend/components/documents`
- `frontend/components/maps`
- `frontend/components/timeline`
- `frontend/pages`
