+++ frontend/components/maps/MapMacro.js
