+++ frontend/index.js
@@ -3,14 +3,14 @@ import ReactDOM from "react-dom";
 import "./scss/index.scss";
 import "bootstrap/dist/css/bootstrap.min.css";
 import Base from "./components/global/Base";
-import ErrorNotFoundComponent from "./components/ErrorNotFoundComponent";
+import ErrorNotFoundComponent from "./components/global/ErrorNotFoundComponent";
 import ExampleId from "./components/ExampleId";
-import Home from "./components/Home";
-import {Timeline} from "./components/Timeline";
-import MapMicro from "./components/MapMicro";
-import MapMacro from "./components/MapMacro";
-import TimelineTest from "./components/TimelineTest";
-import API from "./components/API";
+import Home from "./pages/Home";
+import {Timeline} from "./components/timeline/Timeline";
+import MapMicro from "./components/maps/MapMicro";
+import MapMacro from "./components/maps/MapMacro";
+import TimelineTest from "./components/timeline/TimelineTest";
+import API from "./pages/API";
 
 const COMPONENT_PROPS_RAW = document.getElementById("component_props").text;
 const COMPONENT_NAME_RAW = document.getElementById("component_name").text;
